refactor: relocated auth helper functions from User model to auth service
