+++ alembic/versions/544d9d1c80d3_removed_auth_functions_at_user_model.py
@@ -1,0 +1,30 @@
+"""Removed auth functions at User model
+
+Revision ID: 544d9d1c80d3
+Revises: 901d2fa2b680
+Create Date: 2024-03-02 22:41:32.467176
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '544d9d1c80d3'
+down_revision: Union[str, None] = '901d2fa2b680'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
