+++ src/models/users.py
@@ -1,9 +1,5 @@
 import uuid
-import bcrypt
-from datetime import datetime, timedelta, timezone
-from jose import jwt
 from src.utils.db import Base
-from src.utils.settings import REFRESH_TOKEN_SECRET, ACCESS_TOKEN_SECRET
 from sqlalchemy import (
     Boolean,
     Column,
@@ -36,39 +32,3 @@ class User(Base):
     def __repr__(self):
         """Returns string representation of model instance"""
         return "<User {first_name}>".format(first_name=self.first_name)
-
-    @staticmethod
-    def hash_password(password: bytes) -> bytes:
-        """Transforms password from it's raw textual form to
-        cryptographic hashes
-        """
-        return bcrypt.hashpw(password, bcrypt.gensalt())
-
-    def validate_password(self, password: str) -> bool:
-        """Confirms password validity"""
-        return bcrypt.checkpw(password.encode(), self.hashed_password)
-
-    def generate_token(self):
-        """Generate access token and refresh token for user"""
-        refresh_token = jwt.encode(
-            {
-                "first_name": self.first_name,
-                "email": self.email,
-                "exp": datetime.now(timezone.utc) + timedelta(days=7),
-            },
-            REFRESH_TOKEN_SECRET,
-        )
-        access_token = jwt.encode(
-            {
-                "first_name": self.first_name,
-                "email": self.email,
-                "exp": datetime.now(timezone.utc) + timedelta(minutes=15),
-            },
-            ACCESS_TOKEN_SECRET,
-        )
-        self.refresh_token = refresh_token
-        self.access_token = access_token
-        return {
-            "refresh_token": refresh_token,
-            "access_token": access_token,
-        }
